refactor(core): move everything one more level higher
